clean up key loading
