+++ src/server_certificate.hpp
@@ -1,119 +1,31 @@
-//
-// Copyright (c) 2016-2019 Vinnie Falco (vinnie dot falco at gmail dot com)
-//
-// Distributed under the Boost Software License, Version 1.0. (See accompanying
-// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-//
-// Official repository: https://github.com/boostorg/beast
-//
+/*
+MIT License
+
+Copyright (c) 2019 Chris McArthur, prince.chrismc(at)gmail(dot)com
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
 
 #ifndef BOOST_BEAST_EXAMPLE_COMMON_SERVER_CERTIFICATE_HPP
 #define BOOST_BEAST_EXAMPLE_COMMON_SERVER_CERTIFICATE_HPP
 
-#include <boost/asio/buffer.hpp>
-#include <boost/asio/error.hpp>
-#include <boost/asio/ssl/context.hpp>
-#include <cstddef>
-#include <memory>
-
-BIO* make_buffer_bio(const boost::asio::const_buffer& b)
-{
-  return ::BIO_new_mem_buf(
-      const_cast<void*>(b.data()),
-      static_cast<int>(b.size()));
-}
-
-struct bio_cleanup {
-  BIO* p;
-  ~bio_cleanup()
-  {
-    if (p) ::BIO_free(p);
-  }
-};
-
-struct x509_cleanup {
-  X509* p;
-  ~x509_cleanup()
-  {
-    if (p) ::X509_free(p);
-  }
-};
-
-struct evp_pkey_cleanup {
-  EVP_PKEY* p;
-  ~evp_pkey_cleanup()
-  {
-    if (p) ::EVP_PKEY_free(p);
-  }
-};
-
-struct ec_key_cleanup {
-  EC_KEY* p;
-  ~ec_key_cleanup()
-  {
-    if (p) ::EC_KEY_free(p);
-  }
-};
-
-BOOST_ASIO_SYNC_OP_VOID do_use_tmp_ecdh(
-    boost::asio::ssl::context& ctx,
-    BIO* bio,
-    boost::system::error_code& ec)
-{
-  ::ERR_clear_error();
-
-  int nid = NID_undef;
-
-  x509_cleanup x509 = {::PEM_read_bio_X509(bio, NULL, 0, NULL)};
-  if (x509.p) {
-    evp_pkey_cleanup pkey = {::X509_get_pubkey(x509.p)};
-    if (pkey.p) {
-      ec_key_cleanup tmp = {::EVP_PKEY_get1_EC_KEY(pkey.p)};
-      if (tmp.p) {
-        const EC_GROUP* group = EC_KEY_get0_group(tmp.p);
-        nid = EC_GROUP_get_curve_name(group);
-      }
-    }
-  }
-
-  ec_key_cleanup ec_key = {::EC_KEY_new_by_curve_name(nid)};
-  if (ec_key.p) {
-    if (::SSL_CTX_set_tmp_ecdh(ctx.native_handle(), ec_key.p) == 1) {
-      ec = boost::system::error_code();
-      BOOST_ASIO_SYNC_OP_VOID_RETURN(ec);
-    }
-  }
-
-  ec = boost::system::error_code(
-      static_cast<int>(::ERR_get_error()),
-      boost::asio::error::get_ssl_category());
-  BOOST_ASIO_SYNC_OP_VOID_RETURN(ec);
-}
-
-BOOST_ASIO_SYNC_OP_VOID use_tmp_ecdh_file(boost::asio::ssl::context& ctx,
-                                          const boost::asio::const_buffer& certificate,
-                                          boost::system::error_code& ec)
-{
-  ::ERR_clear_error();
-
-  bio_cleanup bio = {make_buffer_bio(certificate)};
-  if (bio.p) {
-    return do_use_tmp_ecdh(ctx, bio.p, ec);
-  }
-
-  ec = boost::system::error_code(
-      static_cast<int>(::ERR_get_error()),
-      boost::asio::error::get_ssl_category());
-  BOOST_ASIO_SYNC_OP_VOID_RETURN(ec);
-}
-
-void use_tmp_ecdh(boost::asio::ssl::context& ctx,
-                  const boost::asio::const_buffer& certificate)
-{
-  boost::system::error_code ec;
-  use_tmp_ecdh_file(ctx, certificate, ec);
-  boost::asio::detail::throw_error(ec, "use_tmp_ecdh");
-}
+#include "use_tmp_ecdh.hpp"
 
 // Load a signed certificate into the ssl context, and configure
 // the context for use with a server.
@@ -149,7 +61,6 @@ BBgwFoAUszKCfzlSl54nDpRdS1DGnOcAIRAwDwYDVR0TAQH/BAUwAwEB/zAOBgNV
 HQ8BAf8EBAMCAYYwCgYIKoZIzj0EAwIDRwAwRAIgP2nFxq3xx6j21FJS/fgCjKUs
 vD/s8xa95p+ZrN/Zu6ECID1hk+lh4sHtTr+LQunwc43a0l4EF05XvHK8/DwsP660
 -----END CERTIFICATE-----
-
 )###";
 
   const std::string key = R"###(-----BEGIN EC PRIVATE KEY-----
